replace sqlite with postgres

+++ common/nyxd-scraper/sql_migrations/01_metadata.sql
@@ -2,9 +2,8 @@
  * Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
  * SPDX-License-Identifier: Apache-2.0
  */
-
-CREATE TABLE METADATA
-(
-    id                    INTEGER PRIMARY KEY CHECK (id = 0),
-    last_processed_height INTEGER NOT NULL
-);
+CREATE TABLE
+    METADATA (
+        id INTEGER PRIMARY KEY CHECK (id = 0),
+        last_processed_height BIGINT NOT NULL
+    );
